@@ -5,33 +5,21 @@
   "author": "silverwind <me@silverwind.io>",
   "repository": "silverwind/daemonize-process",
   "license": "BSD-2-Clause",
-  "main": "index.js",
-  "scripts": {
-    "test": "make test"
-  },
+  "exports": "./index.js",
+  "type": "module",
+  "sideEffects": false,
   "engines": {
-    "node": ">=10"
+    "node": ">=18"
   },
   "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "daemonize",
-    "daemon",
-    "background",
-    "process"
+    "./index.js"
   ],
   "devDependencies": {
-    "del": "5.1.0",
-    "eslint": "6.8.0",
-    "eslint-config-silverwind": "10.0.1",
-    "jest": "25.1.0",
-    "tempy": "0.5.0",
-    "updates": "10.2.3",
-    "versions": "8.2.5"
-  },
-  "jest": {
-    "verbose": false,
-    "testTimeout": 10000
+    "eslint": "8.49.0",
+    "eslint-config-silverwind": "74.1.4",
+    "updates": "15.0.2",
+    "versions": "11.1.0",
+    "vitest": "0.34.4",
+    "vitest-config-silverwind": "3.0.0"
   }
 }
